requirements.txt
asgiref==3.8.1
Django==5.2.2
paho-mqtt==2.1.0
sqlparse==0.5.3
